(halium) frameworks/av: backport "Camera: skip unnecessary reconfig"
This patch is backported from Android 10, and fixes zooming stutter/
freeze on VollaPhone.

Turns out that "reconfigure" is pretty costly, seemingly because it has
to drain the buffers first. And it happens every time the parameters are
set (which is, every time zooming happens).

Interestingly, after applying the patch, the app doesn't freeze anymore,
but the zooming is delayed. I suspect that delay doesn't present on
Android, which is probably the reason it didn't pose a problem earlier.

Change-Id: I4aa2dea4be4c1ad1fd5ee1c1bfd471d659668db5